+++ scoring/src/fll/web/admin/CommitTeam.java
@@ -80,7 +80,7 @@ public class CommitTeam extends BaseFLLServlet {
     try {
       connection = datasource.getConnection();
       // parse the numbers first so that we don't get a partial commit
-      final int teamNumber = Utilities.NUMBER_FORMAT_INSTANCE.parse(request.getParameter("teamNumber")).intValue();
+      final int teamNumber = Utilities.INTEGER_NUMBER_FORMAT_INSTANCE.parse(request.getParameter("teamNumber")).intValue();
 
       String redirect = null;
       if (null != request.getParameter("delete")) {