+++ Changes.md
@@ -1,3 +1,4 @@
+* #528 - Only display floating point scores when needed, otherwise display integers
 * #566 - Use web sockets for the unverified runs update
 * #552 - use web sockets to update the remote control brackets page 
 * upgrade to selenium 3.3.0